Put orthanc_sdk enums in dedicated file
